requirements.txt
fast-bitrix24
django
djangorestframework
python-dotenv